src/main/java/school/faang/user_service/service/filter/goal/InvitationInviterNameFilter.java
package school.faang.user_service.service.filter.goal;

import org.springframework.stereotype.Component;
import school.faang.user_service.dto.goal.InvitationFilterDto;
import school.faang.user_service.entity.goal.GoalInvitation;
import school.faang.user_service.service.filter.goal.InvitationFilter;

import java.util.stream.Stream;

@Component
public class InvitationInviterNameFilter implements InvitationFilter {
    @Override
    public boolean isApplicable(InvitationFilterDto filterDto) {
        return filterDto.getInviterNamePattern() != null;
    }

    @Override
    public Stream<GoalInvitation> apply(Stream<GoalInvitation> invitations, InvitationFilterDto filterDto) {
        return invitations.filter(invitation ->
                invitation.getInviter().getUsername().contains(filterDto.getInviterNamePattern()));
    }
}
